Introduce a protocol version and a handshake command
The controller may send a CheckPeer command which the ODC plugin replies
to by a PeerState reply. It is meant to be used as a handshake message
very early in the communication between the controller and the ODC
plugin. For now it is used to exchange a protocol version number.